+++ packages/core/src/Reporter.ts
@@ -184,7 +184,7 @@ export default class Reporter<Options = {}> extends Plugin<Options> {
    * Sort all tasks by start time and filter to remove pending tasks.
    */
   sortTasksByStartTime<T extends Task<any>>(tasks: T[]): T[] {
-    return tasks.sort((a, b) => a.metadata.startTime - b.metadata.startTime);
+    return [...tasks].sort((a, b) => a.metadata.startTime - b.metadata.startTime);
   }
 
   /**